RotationAction

+++ starships/src/main/scala/action/movement/ConstantRotationActionBuilder.scala
@@ -2,6 +2,6 @@ package action.movement
 
 import action.movement.{ConstantMovementAction, MovementAction, MovementActionBuilder, MovementActionBuilderInput}
 
-class ConstantRotationActionBuilder(rotationValueInDegrees: Int) extends MovementActionBuilder:
+class ConstantRotationActionBuilder(rotationValueInDegrees: Double) extends MovementActionBuilder:
   def build(actionBuilderInput: MovementActionBuilderInput): MovementAction =
     ConstantRotationAction(rotationValueInDegrees, actionBuilderInput.entity)
